UC4.4

+++ src/main/java/com/se330/coffee_shop_management_backend/dto/request/order/OrderDetailCreateRequestDTO.java
@@ -3,13 +3,14 @@ package com.se330.coffee_shop_management_backend.dto.request.order;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
 @NoArgsConstructor
 public class OrderDetailCreateRequestDTO {
     private int orderDetailQuantity;
-    private int orderDetailUnitPrice;
+    private BigDecimal orderDetailUnitPrice;
     private UUID productVariantId;
     private UUID orderId;
 }
