+++ src/main/java/com/se330/coffee_shop_management_backend/dto/request/order/OrderDetailUpdateRequestDTO.java
@@ -3,6 +3,7 @@ package com.se330.coffee_shop_management_backend.dto.request.order;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -10,7 +11,7 @@ import java.util.UUID;
 public class OrderDetailUpdateRequestDTO {
     private UUID orderDetailId;
     private int orderDetailQuantity;
-    private int orderDetailUnitPrice;
+    private BigDecimal orderDetailUnitPrice;
     private UUID productVariantId;
     private UUID orderId;
 }
